% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_exceedance_prob.R
\name{annual_exceedance_prob}
\alias{annual_exceedance_prob}
\title{Annual Exceedance Probability}
\usage{
annual_exceedance_prob(u10, weib_a, weib_k)
}
\arguments{
\item{u10}{The critical wind speed (m s-1) at anemometer's height.}

\item{weib_a}{The scale parameter of the Weibull distribution of local wind speeds.}

\item{weib_k}{The shape parameter of the Weibull distribution of local wind speeds.}
}
\value{
\code{aep}, i.e. the probability of exceeding the critical wind speed of damage.
}
\description{
Calculates the annual probability of exceeding a critical wind speed.
}
