% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bending_moment_rou.R
\name{bending_moment_rou}
\alias{bending_moment_rou}
\title{Bending moment on a tree under wind loading}
\usage{
bending_moment_rou(dbh, ht, cr_width, cr_depth, spacing, dist_edge, gap_size,
  uguess, n_drag, c_drag, drag_upper_limit)
}
\arguments{
\item{dbh}{The diameter (cm) of the stem measured at breast height.}

\item{ht}{The height (m) of the tree.}

\item{cr_width}{The width (m) of the tree crown.}

\item{cr_depth}{The depth (m) of the tree crown.}

\item{spacing}{The spacing (m) between trees.}

\item{dist_edge}{The distance (m) of the the tree from the upwind edge.}

\item{gap_size}{The length (m) of the upwind gap.}

\item{uguess}{The speed (m s-1) of the wind.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless.}

\item{drag_upper_limit}{The experimental maximum wind speed (m s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}.}
}
\value{
The bending moment (N) exerted on a tree by the applied wind force.
}
\description{
Calculates the bending moment on a tree resulting from the force applied by the wind.
}
