% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_breadth_fun.R
\name{canopy_breadth_fun}
\alias{canopy_breadth_fun}
\title{Canopy Breadth Function.}
\usage{
canopy_breadth_fun(cr_width, uguess, n_drag, c_drag, drag_upper_limit)
}
\arguments{
\item{cr_width}{The width (m) of the canopy in windless conditions.}

\item{uguess}{The speed (m s-1) of the wind.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}.}
}
\value{
The breadth of the crown under wind loading.
}
\description{
Calculates the breadth of the canopy under wind load.
}
