% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_moments.R
\name{Critical_Moments_Functions}
\alias{Critical_Moments_Functions}
\alias{critical_moment_breakage}
\alias{critical_moment_overturning}
\title{Critical Moments Functions}
\usage{
critical_moment_breakage(dbh, ht, cr_depth, mor, fknot)

critical_moment_overturning(c_reg, stem_density, stem_vol)
}
\arguments{
\item{dbh}{The dbh (cm) of a tree.}

\item{ht}{The height (m) of a tree.}

\item{cr_depth}{The length (m) of the tree crown.}

\item{mor}{Modulus of Rupture (MPa) of green wood.}

\item{fknot}{Knot factor. Dimensionless.}

\item{c_reg}{Regression coefficients (N m kg-1) of uprooting moment against stem weight.}

\item{stem_density}{Density (kg m-3) of green wood of the stem.}

\item{stem_vol}{Volume (m3) of the stem of the mean tree in the stand.}
}
\description{
Calculate the critical moments for breakage and overturning
}
