% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deflection_Loading Factor.R
\name{DFL_Functions}
\alias{DFL_Functions}
\alias{dlf_fun}
\alias{deflection_fun}
\alias{force_of_wind_fun}
\alias{r_fun}
\alias{i_fun}
\title{Deflection Loading Factor Functions}
\usage{
dlf_fun(bm, ht, cr_depth, cr_width, stem_vol, dbh, moe, crown_density,
  stem_density, snow_depth, snow_density)

deflection_fun(x, lever_arm, fow, dbh, ht, cr_depth, pull_height, moe)

force_of_wind_fun(bm, ht, cr_depth)

r_fun(lever_arm, pull_height)

i_fun(dbh, ht, cr_depth)
}
\arguments{
\item{bm}{The applied bending moment resulting from the wind load}

\item{ht}{The height (m) of a tree.}

\item{cr_depth}{The length (m) of the tree crown.}

\item{cr_width}{The width (m) of the tree crown.}

\item{stem_vol}{Volume (m3) of the stem of the mean tree in the stand.}

\item{dbh}{The dbh (cm) of a tree.}

\item{moe}{Modulus of Elasticity (MPa) of green wood.}

\item{crown_density}{Density (kgm-3) of of the crown of the mean tree in the stand.}

\item{stem_density}{Density (kg m-3) of green wood of the stem.}

\item{snow_depth}{Depth (cm) of layer of snow on tree crown.}

\item{snow_density}{Density (kg m-3) of snow.}

\item{x}{Height (m) along the tree stem where deflection is to be calculated.}

\item{lever_arm}{Length (m) of the lever arm. Typically, it is equal tree height.}

\item{fow}{Applied wind loading. In the \code{deflection_fun} function, it is a placeholder for the output of the \code{force_of_wind_fun} function.}

\item{pull_height}{Height (m) along the tree stem where the wind loading is applied.}

\item{stem_density}{Density (kg m-3) of green wood of the stem.}
}
\description{
Calculate the Deflection Loading Factor to account for the additional moment provided by the weight of the stem, crown (and snow, when present)
}
