% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drag_fun.R
\name{drag_fun}
\alias{drag_fun}
\title{Crown Drag.}
\usage{
drag_fun(uguess, n_drag, c_drag, drag_upper_limit)
}
\arguments{
\item{uguess}{Critical wind speed at canopy top calculated with the roughness or single-tree method (m s-1).}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}
}
\value{
The porosity of the crown under wind loading.
}
\description{
Calculates the drag of the canopy under wind load.
}
