% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_gap_factor.R
\name{edge_gap_factor_fun}
\alias{edge_gap_factor_fun}
\title{Edge Gap Factor.}
\usage{
edge_gap_factor_fun(spacing, ht, dist_edge, gap_size)
}
\arguments{
\item{spacing}{The distance (m) between trees.}

\item{ht}{The height of the tree.}

\item{dist_edge}{The distance (m) of the tree from the upwind edge.}

\item{gap_size}{The length (m) of the upwind gap.}
}
\value{
\code{edge_gap_factor}, the combined effect of edge and gap on the maximum bending moment exerted by the wind on a tree. Used in the tmc method.
}
\description{
Combined effect of gap size and edge effect on maximum bending moment for the single-tree method.
}
