% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevate.R
\name{elevate}
\alias{elevate}
\title{Elevate}
\usage{
elevate(uh, z0, d, ht)
}
\arguments{
\item{uh}{The critical wind speed (m s-1) calculated with the roughness or single-tree method.}

\item{z0}{The roughness (m) of the canopy.}

\item{d}{The zero plane displacement height (m).}

\item{ht}{The height of the tree.}
}
\value{
A critical wind speed tht can be compared with met data.
}
\description{
Converts the critical wind speeds to the wind speed 10 meters above zero plane displacement (for anemometers' data). The wind is assumed to follow a logarithmic profile.
}
