% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_rou.R
\name{fg_rou}
\alias{fg_rou}
\title{ForestGALES - Roughness method.}
\usage{
fg_rou(stand_id, date = NA, species, mean_ht, mean_dbh, spacing,
  full_output = 1, weib_a = NA, weib_k = NA, top_ht = NA,
  elev_ht = NA, mean_cr_width = NA, mean_cr_depth = NA,
  soil_group = NA, rooting = NA, new_edge = NA, gap_size = NA,
  species_parameters = fgr:::species_parameters,
  fgr_constants = fgr:::fgr_constants, moe = NA, mor = NA,
  fknot = NA, stem_vol = NA, crown_vol = NA, stem_density = NA,
  crown_density = NA, c_reg = NA, c_drag = NA, n_drag = NA,
  drag_upper_limit = NA, snow_depth = NA, snow_density = NA,
  ro = NA, x = NA, y = NA, z = NA, dams = NA)
}
\arguments{
\item{stand_id}{Stand identifier. Essential.}

\item{date}{Date of wind damage risk assessment. Optional (but useful! Mind the variable order in the function call!).}

\item{species}{Tree species under investigation. Essential.}

\item{mean_ht}{Arithmetic mean height of the trees in the stand (m). Essential (unless \code{top_ht} is provided).}

\item{mean_dbh}{Mean dbh of all trees in the stand (cm). Dbh is diameter at breast height, measured at 1.3m above the ground. Essential.}

\item{spacing}{Mean distance between trees (m). Essential.}

\item{full_output}{Switch between full and basic outputs.}

\item{weib_a}{Scale parameter of the Weibull distribution for local wind speeds.}

\item{weib_k}{Shape parameter of the Weibull distribution for local wind speeds.}

\item{top_ht}{The height of the dominant tree(s) in the stand (m). Essential (unless \code{mean_ht} is provided).}

\item{elev_ht}{The height at which the elevated critical wind speed is calculated (m). For the roughness method, this defaults to mean tree height}

\item{mean_cr_width}{Width of the crown of the mean tree in the stand (m).}

\item{mean_cr_depth}{Length of the crown of the mean tree in the stand (m).}

\item{soil_group}{Soil group identifier (1 = freely draining mineral soils; 2 = gleyed mineral soils; 3 = peaty mineral soils; 4 = deep peats).}

\item{rooting}{Rooting depth class (1 = Shallow, 2 = Deep, 3 = Medium).}

\item{new_edge}{Switch to toggle between green upwind edge (0) or brown upwind edge (1)}

\item{gap_size}{Length of the upwind gap (m).}

\item{species_parameters}{Data.frame covering all species-specific parameters. Defaults to fgr Internal Data species paramters data.frame. If specified by the user, it can have as many or as few species (i.e. rows) as required. It is recommended that when working with user-defined species, stem volume is specified and provided as an input (see below in input list). When this is not possible, stem volume is calculated with Eq. 4 of Fonweban et al. 2012. In order to allow this, letter codes for User-defined species must begin with a capital "U" for the model to recognise that the Fonweban et al. formula needs applied to calculate stem volume.}

\item{fgr_constants}{Data.frame including all constants required for model execution. Defaults to fgr Internal Data fgr constants data.frame. these include entities that can be individually supplied to the model (e.g. ro, the air density, and snow density - see below in input list), or entities that qould not normally require modification (e.g. constants used in the calculation of the Deflection Loading Factor, the Annual Exceedance Probability, the Edge, Gap, and Gust Factors, aerodynamic quantities such as zpd, z0, and gammasolved - but even von Karman's and the gravitational constants can be modified!)}

\item{moe}{Modulus of Elasticity of green wood (MPa). Advanced Input.}

\item{mor}{Modulus of Rupture of green wood (MPa). Advanced Input.}

\item{fknot}{Knot factor (dimensionless). Advanced Input.}

\item{stem_vol}{Stem volume of the mean tree in the stand (m^3). Advanced Input.}

\item{crown_vol}{Volume of the crown of the mean tree (m^3). Advanced Input.}

\item{stem_density}{Density of green wood of the stem (kg m-3). Advanced Input.}

\item{crown_density}{Density of the crown of the mean tree (kg m-3). Advanced Input.}

\item{c_reg}{Regression coefficients of uprooting moment against stem weight (N m kg-1). Advanced Input.}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless). Advanced Input.}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless). Advanced Input.}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1). Advanced Input.}

\item{snow_depth}{Depth of layer of snow on tree crown (cm). Advanced Input.}

\item{snow_density}{Density of snow (kg m-3). Advanced Input.}

\item{ro}{Air density (kg m-3). Advanced Input.}

\item{x}{Spatial coordinate.}

\item{y}{Spatial coordinate.}

\item{z}{Spatial coordinate.}

\item{dams}{DAMS (Detailed Aspect Method of Scoring) value to describe the local wind climate (for UK conditions only).}
}
\value{
If \code{full_output} = 1, a comprehensive list including: stand id; for breakage and overturning, the critical wind speeds (m s-1) at
 canopy top height and anemometer's height; zero plane displacement height (m), canopy surface roughness (m); canopy drag; ratio of critical wind speed at
 canopy top over friction velocity (uh / u*); drag per unit ground area; the deflection loading factor; the critical breaking moment and the critical
 overturning moment; the combined effect of edge, gap, and gustiness on the applied bending moment; the probabilities of damage; the mode of damage;
 a summary of the inputs. If \code{full_output} = 0, a much shorter list including: stand id; the critical wind speeds of breakage and overturning,
 and the associated probabilities.
}
\description{
Wrapper for the roughness method.
}
