% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_tmc.R
\name{fg_tmc}
\alias{fg_tmc}
\title{ForestGALES - Single-tree (TMC) method.}
\usage{
fg_tmc(stand_id, tree_id, species, tree_ht, dbh, spacing_current,
  predominant_species, stand_mean_ht, stand_top_ht, stand_mean_dbh,
  full_output = 1, weib_a = NA, weib_k = NA, ci = NA, ci_value = NA,
  cr_width = NA, cr_depth = NA, stand_cr_width = NA,
  stand_cr_depth = NA, soil_group = NA, rooting = NA, dist_edge = NA,
  gap_size = NA, spacing_before = NA, years_since_thin = NA, moe = NA,
  mor = NA, fknot = NA, stem_vol = NA, crown_vol = NA,
  stem_density = NA, crown_density = NA, c_reg = NA, c_drag = NA,
  n_drag = NA, drag_upper_limit = NA, snow_depth = NA,
  snow_density = NA, ro = NA, x = NA, y = NA, z = NA, dams = NA)
}
\arguments{
\item{stand_id}{The stand identifier. Essential.}

\item{tree_id}{The tree identifier. Essential.}

\item{species}{The tree species under investigation. Essential.}

\item{tree_ht}{The height (m) of a tree. Essential.}

\item{dbh}{The dbh (cm) of a tree. Essential.}

\item{spacing_current}{The current mean spacing (m) between trees in the stand. Essential.}

\item{predominant_species}{The predominant species in the stand.}

\item{stand_mean_ht}{Height (m) of the mean tree in the stand. Essential (unless \code{stand_mean_ht} is provided).}

\item{stand_top_ht}{Height (m) of the dominant tree(s) in the stand. Essential (unless \code{stand_top_ht} is provided).}

\item{stand_mean_dbh}{Dbh (cm) of the dominant tree(s) in the stand. Essential.}

\item{full_output}{Switch between full and minimal outputs.}

\item{weib_a}{The scale parameter of the Weibull distribution of local wind speeds.}

\item{weib_k}{The shape parameter of the Weibull distribution of local wind speeds.}

\item{ci}{Competition Index (\code{BAL}, \code{Hegyi}) used. Can be \code{None}.}

\item{ci_value}{Value of \code{ci}.}

\item{cr_width}{The width (m) of the tree crown.}

\item{cr_depth}{The length (m) of the tree crown.}

\item{stand_cr_width}{The width (m) of the crown of the mean tree in the stand.}

\item{stand_cr_depth}{The length (m) of the crown of the mean tree in the stand.}

\item{soil_group}{The soil type identifier.}

\item{rooting}{The rooting depth class (1 = Shallow, 2 = Deep, 3 = Medium).}

\item{dist_edge}{Distance (m) of the mean tree from the upwind edge.}

\item{gap_size}{Length (m) of the upwind gap.}

\item{spacing_before}{The mean spacing (m) between trees in the stand before any thinning.}

\item{years_since_thin}{The number of years after the latest thinning.}

\item{moe}{Modulus of Elasticity (MPa) of green wood. Advanced Input.}

\item{mor}{Modulus of Rupture (MPa) of green wood. Advanced Input.}

\item{fknot}{Knot factor. Dimensionless. Advanced Input.}

\item{stem_vol}{Volume (m3) of the stem of the tree. Advanced Input.}

\item{crown_vol}{Volume (m3) of the tree crown. Advanced Input.}

\item{stem_density}{Density (kg m-3) of green wood of the stem. Advanced Input.}

\item{crown_density}{Density (kg m-3) of of the crown of the tree. Advanced Input.}

\item{c_reg}{Regression coefficients (N m kg-1) of uprooting moment against stem weight. Advanced Input.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless. Advanced Input.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless. Advanced Input.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}. Advanced Input.}

\item{snow_depth}{Depth (cm) of layer of snow on tree crown. Advanced Input.}

\item{snow_density}{Density (kg m-3) of snow. Advanced Input.}

\item{ro}{Air density (kg m-3). Advanced Input.}

\item{x}{Spatial coordinate.}

\item{y}{Spatial coordinate.}

\item{z}{Spatial coordinate.}

\item{dams}{Value of Detailed Aspect Method of Scoring to describe the local wind climate (for UK conditions only).}

\item{status}{The dominance class of the tree under investigation.}
}
\value{
If \code{full_output} = 1, a comprehensive list including: stand and tree id's; for breakage and overturning, the critical wind speeds (m s-1) at
 canopy top height and anemometer's height; zero plane displacement height (m), canopy surface roughness (m); canopy drag; ratio of critical wind speed at
 canopy top over friction velocity (uh / u*); drag per unit ground area; the turning moment coefficient; the turning moment ratio; the deflection loading
 factor; the critical breaking moment and the critical overturning moment; the combined effect of edge and gap on the applied bending moment; the probabilities
 of damage; the mode of damage; a summary of the inputs. If \code{full_output} = 0, a much shorter list including: stand id and tree id's; the critical
 wind speeds of breakage and overturning, and the associated probabilities.
}
\description{
Wrapper for the single-tree (turning moment coefficient) method.
}
