% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammasolved_fun.R
\name{gammasolved_fun}
\alias{gammasolved_fun}
\title{Gamma Solved Function.}
\usage{
gammasolved_fun(cr_width, cr_depth, spacing, uguess, n_drag, c_drag,
  drag_upper_limit)
}
\arguments{
\item{cr_width}{The width (m) of the canopy in windless conditions.}

\item{cr_depth}{The length (m) of the canopy in windless conditions.}

\item{spacing}{The spacing (m) between trees.}

\item{uguess}{The speed (m s-1) of the wind.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}.}
}
\value{
\code{gammasolved}, the ratio of critical wind speed at canopy top over friction velocity (uh / u*).
}
\description{
Calculates the ratio between the critical wind speed at canopy top and the friction velocity (uh / u*).
}
