% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_vol_switch.R
\name{stem_vol_fun}
\alias{stem_vol_fun}
\title{Stem Volume Function.}
\usage{
stem_vol_fun(species, dbh, ht, species_parameters)
}
\arguments{
\item{species}{Tree species. If working with a non-default species (i.e. named "U^"), the stem_volume_fonweban function to calulate volume is adopted.}

\item{dbh}{Diameter of the stem at breast height, i.e. 1.3m above the ground (cm). Depending on the method used ('roughness' or 'TMC') this can be either the arithmetic average of the dbh of all the trees in the stand, or the dbh of an individual tree.}

\item{ht}{Tree height. Depending on the method used ('roughness' or 'TMC'), this can be either the mean tree in the stand, or each individual tree (m).}
}
\value{
The volume of the stem (m3).
}
\description{
Acts as a switch to select the appropriate function to calculate the stem volume of a tree.
}
