% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws_tmc_breakage.R
\name{TMC_Breakage_Critical_wind_speed_functions}
\alias{TMC_Breakage_Critical_wind_speed_functions}
\alias{uh_breakage_tmc}
\alias{uh_breakage_tmc_tmr_simple}
\title{Critical wind speed for breakage - Single-tree method.}
\usage{
uh_breakage_tmc(
  tree_ht,
  dbh,
  cr_depth,
  cr_width,
  spacing_current,
  spacing_before,
  years_since_thin,
  dist_edge,
  gap_size,
  moe,
  mor,
  fknot,
  stem_vol,
  stem_density,
  crown_density,
  snow_depth,
  snow_density,
  ci,
  ci_value,
  n_drag,
  c_drag,
  drag_upper_limit,
  equivalent_mean_ht,
  stand_cr_depth,
  stand_cr_width,
  fgr_constants
)

uh_breakage_tmc_tmr_simple(
  tree_ht,
  dbh,
  cr_depth,
  cr_width,
  spacing_current,
  spacing_before,
  years_since_thin,
  dist_edge,
  gap_size,
  equivalent_mean_ht,
  moe,
  mor,
  fknot,
  stem_vol,
  stem_density,
  crown_density,
  snow_depth,
  snow_density,
  ci,
  ci_value,
  fgr_constants
)
}
\arguments{
\item{tree_ht}{Individual tree height (m).}

\item{dbh}{Diameter of the stem at breast height, i.e. 1.3m above the ground (cm).}

\item{cr_depth}{Length of the tree crown (m).}

\item{cr_width}{Width of the tree crown (m).}

\item{spacing_current}{Current mean spacing of trees in the stand (m).}

\item{spacing_before}{Mean spacing of trees in the stand before any thinning (m).}

\item{years_since_thin}{Number of years since most recent thinning.}

\item{dist_edge}{Distance of tree from the upwind edge (m).}

\item{gap_size}{Length of the upwind gap (x).}

\item{moe}{Modulus of Elasticity of green wood (MPa).}

\item{mor}{Modulus of Rupture of green wood (MPa).}

\item{fknot}{Knot factor (dimensionless).}

\item{stem_vol}{Individual tree stem volume (m3).}

\item{stem_density}{Density of green wood of the stem (kg m-3).}

\item{crown_density}{Density of of the tree crown (kg m-3).}

\item{snow_depth}{Depth of layer of snow on tree crown (cm).}

\item{snow_density}{Density of snow (kg m-3).}

\item{ci}{Competition Index (\code{bal}, \code{heg}, \code{none}) used.}

\item{ci_value}{Value of \code{ci}.}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}

\item{equivalent_mean_ht}{Equivalent mean stand height: the level in the stand responsible for most of the momentum absorption (m).}

\item{stand_cr_depth}{Length of the crown of the "equivalent mean tree" in the stand (m).}

\item{stand_cr_width}{Width of the crown of the "equivalent mean tree" in the stand (m).}

\item{stand_mean_ht}{Arithmetic mean height of the trees in the stand (m).}

\item{c_reg}{Regression coefficients of uprooting moment against stem weight (N m kg-1).}
}
\value{
A list including: The critical wind speed (m s-1) for breakage at canopy top height for the single-tree method;
The turning moment coefficient; The turning moment ratio; The deflection loading factor; The critical breaking moment;
The combined effect of edge and gap on the applied bending moment.
}
\description{
Calculates the critical wind speed for breakage using the TMC method.
}
