% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turning_Moment_Ratios.R
\name{Turning_Moment_Ratios}
\alias{Turning_Moment_Ratios}
\alias{tm_ratio_simple}
\alias{tm_ratio_simple_ci}
\alias{tm_ratio}
\alias{tm_ratio_ci}
\title{Turning Moment Ratio Functions}
\usage{
tm_ratio_simple(spacing_before, spacing_current, years_since_thin)

tm_ratio_simple_ci(spacing_before, spacing_current, years_since_thin, ci)

tm_ratio(spacing_before, spacing_current, years_since_thin, cr_width, cr_depth,
  uguess, n_drag, c_drag, drag_upper_limit, ht)

tm_ratio_ci(spacing_before, spacing_current, years_since_thin, cr_width,
  cr_depth, uguess, n_drag, c_drag, drag_upper_limit, ht, ci)
}
\arguments{
\item{spacing_before}{The mean spacing (m) between trees in the stand before any thinning.}

\item{spacing_current}{The current mean spacing (m) between trees in the stand.}

\item{years_since_thin}{The number of years after the latest thinning.}

\item{ci}{Competition Index (\code{BAL}, \code{Hegyi}) used. Can be \code{None}.}

\item{cr_width}{The width (m) of the tree crown.}

\item{cr_depth}{The length (m) of the tree crown.}

\item{uguess}{The speed (m s-1) of the wind.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}.}

\item{ht}{The height (m) of the tree.}
}
\description{
Calculate the ratio between the turning moment coefficient before and after thinning in the stand.
}
