% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws_rou_breakage.R
\name{uh_breakage_rou}
\alias{uh_breakage_rou}
\title{Critical wind speed for breakage - Roughness method.}
\usage{
uh_breakage_rou(mean_ht, mean_dbh, spacing, dist_edge, gap_size, mean_cr_width,
  mean_cr_depth, moe, mor, fknot, n_drag, c_drag, drag_upper_limit, stem_vol,
  stem_density, crown_density, snow_depth, snow_density, ro)
}
\arguments{
\item{mean_ht}{The height (m) of the mean tree in the stand.}

\item{mean_dbh}{The dbh (cm) of the mean tree in the stand.}

\item{spacing}{The mean spacing (m) between trees in the stand.}

\item{dist_edge}{Distance (m) of the mean tree from the upwind edge.}

\item{gap_size}{Length (m) of the upwind gap.}

\item{mean_cr_width}{Width (m) of the crown of the mean tree in the stand.}

\item{mean_cr_depth}{Depth (m) of the crown of the mean tree in the stand.}

\item{moe}{Modulus of Elasticity (MPa) of green wood.}

\item{mor}{Modulus of Rupture (MPa) of green wood.}

\item{fknot}{Knot factor. Dimensionless.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}.}

\item{stem_vol}{Volume (m3) of the stem of the mean tree in the stand.}

\item{stem_density}{Density (kg m-3) of green wood of the stem.}

\item{crown_density}{Density (kg m-3) of of the crown of the mean tree in the stand.}

\item{snow_depth}{Depth (cm) of layer of snow on tree crown.}

\item{snow_density}{Density (kg m-3) of snow.}

\item{ro}{Air density (kg m-3).}
}
\value{
A list including: The critical wind speed (m s-1) for breakage at zero plane displacement height for the roughness method;
The applied bending moment; The zero plane displacement height; The ratio of critical wind speed at canopy top over friction velocity
(uh / u*); The deflection loading factor; The critical breaking moment; The combined effect of edge, gap, and gust on the applied
bending moment.
}
\description{
Calculates the critical wind speed for breakage using the roughness method.
}
