% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws_tmc_breakage.R
\name{uh_breakage_tmc}
\alias{uh_breakage_tmc}
\title{Critical wind speed for breakage - Single-tree method.}
\usage{
uh_breakage_tmc(tree_ht, dbh, cr_depth, cr_width, spacing_current,
  spacing_before, years_since_thin, dist_edge, gap_size, stand_mean_ht, moe,
  mor, fknot, stem_vol, stem_density, crown_density, snow_depth, snow_density,
  ci, ci_value)
}
\arguments{
\item{tree_ht}{The height (m) of a tree.}

\item{dbh}{The dbh (cm) of a tree.}

\item{cr_depth}{The length (m) of the tree crown.}

\item{cr_width}{The width (m) of the tree crown.}

\item{spacing_current}{The current mean spacing (m) between trees in the stand.}

\item{spacing_before}{The mean spacing (m) between trees in the stand before any thinning.}

\item{years_since_thin}{The number of years after the latest thinning.}

\item{dist_edge}{Distance (m) of the mean tree from the upwind edge.}

\item{gap_size}{Length (m) of the upwind gap.}

\item{stand_mean_ht}{Height (m) of the mean tree in the stand.}

\item{moe}{Modulus of Elasticity (MPa) of green wood.}

\item{mor}{Modulus of Rupture (MPa) of green wood.}

\item{fknot}{Knot factor. Dimensionless.}

\item{stem_vol}{Volume (m3) of the stem of the tree.}

\item{stem_density}{Density (kg m-3) of green wood of the stem.}

\item{crown_density}{Density (kg m-3) of of the crown of the tree.}

\item{snow_depth}{Depth (cm) of layer of snow on tree crown.}

\item{snow_density}{Density (kg m-3) of snow.}

\item{ci}{Competition Index (\code{BAL}, \code{Hegyi}) used. Can be \code{None}.}

\item{ci_value}{Value of \code{ci}.}
}
\value{
A list including: The critical wind speed (m s-1) for breakage at canopy top height for the single-tree method;
The turning moment coefficient; The turning moment ratio; The deflection loading factor; The critical breaking moment;
The combined effect of edge and gap on the applied bending moment.
}
\description{
Calculates the critical wind speed for breakage using the TMC method.
}
