% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z0_fun.R
\name{z0_fun}
\alias{z0_fun}
\title{Canopy Roughness Function.}
\usage{
z0_fun(cr_width, cr_depth, spacing, uguess, n_drag, c_drag,
  drag_upper_limit, ht, fgr_constants)
}
\arguments{
\item{cr_width}{Width of the canopy in windless conditions (m).}

\item{cr_depth}{Length of the canopy in windless conditions (m).}

\item{spacing}{Mean distance between trees (m).}

\item{uguess}{Critical wind speed at canopy top calculated with the roughness or single-tree method (m s-1).}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}

\item{ht}{Height of the tree. In the 'roughness' method, this is stand mean height. In the TMC method, this is 'equivalent mean height' (m).}
}
\value{
\code{z0}, the length (m) of the surface roughness of the tree canopy.
}
\description{
Calculates the length (m) of the canopy surface roughness.
}
