% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_exceedance_prob.R
\name{annual_exceedance_prob}
\alias{annual_exceedance_prob}
\title{Annual Exceedance Probability}
\usage{
annual_exceedance_prob(u10, weib_a, weib_k, fgr_constants)
}
\arguments{
\item{u10}{Critical wind speed at standard anemometer height, i.e. 10m above zero-plane displacement height (m s-1).}

\item{weib_a}{Scale parameter of the Weibull distribution for local wind speeds.}

\item{weib_k}{Shape parameter of the Weibull distribution for local wind speeds.}
}
\value{
\code{aep}, i.e. the probability of exceeding the critical wind speed of damage.
}
\description{
Calculates the annual probability of exceeding a critical wind speed.
}
