% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_breadth_fun.R
\name{canopy_breadth_fun}
\alias{canopy_breadth_fun}
\title{Canopy Breadth Function.}
\usage{
canopy_breadth_fun(cr_width, uguess, n_drag, c_drag, drag_upper_limit)
}
\arguments{
\item{cr_width}{The width of the canopy in windless conditions (m).}

\item{uguess}{Critical wind speed at canopy top calculated with the roughness or single-tree method (m s-1).}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}
}
\value{
The breadth of the crown under wind loading (m).
}
\description{
Calculates the breadth of the canopy under wind load.
}
