% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deflection_Loading Factor.R
\name{DFL_Functions}
\alias{DFL_Functions}
\alias{dlf_fun}
\alias{deflection_fun}
\alias{force_of_wind_fun}
\alias{r_fun}
\alias{i_fun}
\title{Deflection Loading Factor Functions}
\usage{
dlf_fun(bm, ht, cr_depth, cr_width, stem_vol, dbh, moe, crown_density,
  stem_density, snow_depth, snow_density)

deflection_fun(x, lever_arm, fow, dbh, ht, cr_depth, pull_height, moe)

force_of_wind_fun(bm, ht, cr_depth)

r_fun(lever_arm, pull_height)

i_fun(dbh, ht, cr_depth)
}
\arguments{
\item{bm}{Applied bending moment resulting from the wind load (N).}

\item{ht}{Tree height. Depending on the method used ('roughness' or 'TMC'), this can be either the mean tree in the stand, or each individual tree (m).}

\item{cr_depth}{Length of the tree crown (m).}

\item{cr_width}{Width of the tree crown (m).}

\item{stem_vol}{Volume of the tree stem of the mean tree in the stand (m3). For the roughness method, this is stem volume of the mean tree. For the TMC method, this is individual tree stem voume.}

\item{dbh}{Diameter of the stem at breast height, i.e. 1.3m above the ground (cm). Depending on the method used ('roughness' or 'TMC') this can be either the arithmetic average of the dbh of all the trees in the stand, or the dbh of an individual tree.}

\item{moe}{Modulus of Elasticity of green wood (MPa).}

\item{crown_density}{Density of of the tree crown (kg m-3).}

\item{stem_density}{Density of green wood of the stem (kg m-3).}

\item{snow_depth}{Depth of layer of snow on tree crown (cm).}

\item{snow_density}{Density of snow (kg m-3).}

\item{x}{Height along the tree stem where deflection is to be calculated (m).}

\item{lever_arm}{Length of the lever arm (m). Typically, it is equal tree height.}

\item{fow}{Applied wind loading in the \code{deflection_fun} function. It is a placeholder for the output of the \code{force_of_wind_fun} function.}

\item{pull_height}{Height along the tree stem where the wind loading is applied (m).}

\item{stem_density}{Density of green wood of the stem (kg m-3).}
}
\description{
Calculate the Deflection Loading Factor to account for the additional moment provided by the weight of the stem, crown (and snow, when present)
}
