% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drag_fun.R
\name{drag_fun}
\alias{drag_fun}
\title{Crown Drag.}
\usage{
drag_fun(uguess, n_drag, c_drag, drag_upper_limit)
}
\arguments{
\item{uguess}{The speed (m s-1) of the wind.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}.}
}
\value{
The porosity of the crown under wind loading.
}
\description{
Calculates the drag of the canopy under wind load.
}
