% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_gap_factor.R
\name{edge_gap_factor_fun}
\alias{edge_gap_factor_fun}
\title{Edge Gap Factor.}
\usage{
edge_gap_factor_fun(spacing, equivalent_mean_ht, dist_edge, gap_size,
  fgr_constants)
}
\arguments{
\item{spacing}{Mean distance between trees (m).}

\item{equivalent_mean_ht}{Equivalent mean stand height: the level in the stand responsible for most of the momentum absorption (m).}

\item{dist_edge}{Distance of tree from the upwind edge (m).}

\item{gap_size}{Length of the upwind gap (m).}
}
\value{
\code{edge_gap_factor}, the combined effect of edge and gap on the maximum bending moment exerted by the wind on a tree. Used in the tmc method.
}
\description{
Combined effect of gap size and edge effect on maximum bending moment for the single-tree method.
}
