% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_gap_gust_factor.R
\name{edge_gap_gust_factor_fun}
\alias{edge_gap_gust_factor_fun}
\title{Edge Gap Gust Factor.}
\usage{
edge_gap_gust_factor_fun(spacing, mean_ht, dist_edge, gap_size,
  fgr_constants)
}
\arguments{
\item{spacing}{Mean distance between trees (m).}

\item{mean_ht}{The mean tree height in the stand (m).}

\item{dist_edge}{Distance of tree from the upwind edge (m).}

\item{gap_size}{Length of the upwind gap (m).}
}
\value{
\code{edge_gap_gust_factor}, the combined effect of edge, gap, and gustiness on the maximum bending moment exerted by the wind on a tree. Used in the roughness method.
}
\description{
Combined effect of gap size, edge effect, and gustiness on maximum bending moment for the roughness method.
}
