% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevate.R
\name{elevate}
\alias{elevate}
\title{Elevate}
\usage{
elevate(uh, z0, d, ht)
}
\arguments{
\item{uh}{Critical wind speed at canopy top calculated with the roughness or single-tree method (m s-1).}

\item{z0}{Aerodynamic roughness length of the canopy (m).}

\item{d}{Zero plane displacement height of the canopy (m).}

\item{ht}{Height at which the critical wind speed is calculated (m). Note that this differs between the 'roughness' (stand mean height) and 'TMC' (1.05 * stand top height) methods.}
}
\value{
A critical wind speed that can be compared with met data.
}
\description{
Converts the critical wind speeds to the wind speed 10 meters above zero plane displacement (for anemometers' data). The wind is assumed to follow a logarithmic profile.
}
