% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_rou.R
\name{fg_rou}
\alias{fg_rou}
\title{ForestGALES - Roughness method.}
\usage{
fg_rou(stand_id, species, mean_ht, mean_dbh, spacing, full_output = 1,
  weib_a = NA, weib_k = NA, top_ht = NA, mean_cr_width = NA,
  mean_cr_depth = NA, soil_group = NA, rooting = NA, new_edge = NA,
  gap_size = NA, moe = NA, mor = NA, fknot = NA, stem_vol = NA,
  crown_vol = NA, stem_density = NA, crown_density = NA, c_reg = NA,
  c_drag = NA, n_drag = NA, drag_upper_limit = NA, snow_depth = NA,
  snow_density = NA, ro = NA, x = NA, y = NA, z = NA, dams = NA)
}
\arguments{
\item{stand_id}{The stand identifier. Essential.}

\item{species}{The tree species under investigation. Essential.}

\item{mean_ht}{The mean tree height (m) in the stand. Essential (unless \code{top_ht} is provided).}

\item{mean_dbh}{The mean dbh (cm) of trees in the stand. Essential.}

\item{spacing}{The current mean spacing (m) between trees in the stand. Essential.}

\item{full_output}{Switch between full and minimal outputs.}

\item{weib_a}{The scale parameter of the Weibull distribution of local wind speeds.}

\item{weib_k}{The shape parameter of the Weibull distribution of local wind speeds.}

\item{top_ht}{The height (m) of the dominant tree(s) in the stand. Essential (unless \code{mean_ht} is provided).}

\item{mean_cr_width}{The width (m) of the crown of the mean tree in the stand.}

\item{mean_cr_depth}{The length (m) of the crown of the mean tree in the stand.}

\item{soil_group}{The soil type identifier.}

\item{rooting}{The rooting depth class (1 = Shallow, 2 = Deep, 3 = Medium).}

\item{new_edge}{Switch to toggle between green upwind edge (0) or brown upwind edge (1)}

\item{gap_size}{Length (m) of the upwind gap.}

\item{moe}{Modulus of Elasticity (MPa) of green wood. Advanced Input.}

\item{mor}{Modulus of Rupture (MPa) of green wood. Advanced Input.}

\item{fknot}{Knot factor. Dimensionless. Advanced Input.}

\item{stem_vol}{Volume (m3) of the stem of the tree. Advanced Input.}

\item{crown_vol}{Volume (m3) of the tree crown. Advanced Input.}

\item{stem_density}{Density (kg m-3) of green wood of the stem. Advanced Input.}

\item{crown_density}{Density (kg m-3) of of the crown of the tree. Advanced Input.}

\item{c_reg}{Regression coefficients (N m kg-1) of uprooting moment against stem weight. Advanced Input.}

\item{c_drag}{The C parameter of the drag coefficient formula. Dimensionless. Advanced Input.}

\item{n_drag}{The N parameter of the drag coefficient formula. Dimensionless. Advanced Input.}

\item{drag_upper_limit}{The experimental maximum wind speed (m*s-1) for calculations of the vaules of \code{n_drag} and \code{c_drag}. Advanced Input.}

\item{snow_depth}{Depth (cm) of layer of snow on tree crown. Advanced Input.}

\item{snow_density}{Density (kg m-3) of snow. Advanced Input.}

\item{ro}{Air density (kg m-3). Advanced Input.}

\item{x}{Spatial coordinate.}

\item{y}{Spatial coordinate.}

\item{z}{Spatial coordinate.}

\item{dams}{Value of Detailed Aspect Method of Scoring to describe the local wind climate (for UK conditions only).}
}
\value{
If \code{full_output} = 1, a comprehensive list including: stand id; for breakage and overturning, the critical wind speeds (m s-1) at
 canopy top height and anemometer's height; zero plane displacement height (m), canopy surface roughness (m); canopy drag; ratio of critical wind speed at
 canopy top over friction velocity (uh / u*); drag per unit ground area; the deflection loading factor; the critical breaking moment and the critical
 overturning moment; the combined effect of edge, gap, and gustiness on the applied bending moment; the probabilities of damage; the mode of damage;
 a summary of the inputs. If \code{full_output} = 0, a much shorter list including: stand id; the critical wind speeds of breakage and overturning,
 and the associated probabilities.
}
\description{
Wrapper for the roughness method.
}
