% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_tmc.R
\name{fg_tmc}
\alias{fg_tmc}
\title{ForestGALES - Single-tree (TMC) method.}
\usage{
fg_tmc(
  stand_id,
  tree_id,
  date = NA,
  species,
  tree_ht,
  dbh,
  spacing_current,
  predominant_species = species,
  stand_mean_ht,
  stand_mean_dbh,
  stand_top_ht,
  full_output = 1,
  weib_a = NA,
  weib_k = NA,
  ci = NA,
  ci_value = NA,
  elev_ht = NA,
  cr_width = NA,
  cr_depth = NA,
  stand_cr_width = NA,
  stand_cr_depth = NA,
  soil_group = NA,
  rooting = NA,
  dist_edge = NA,
  gap_size = NA,
  spacing_before = NA,
  years_since_thin = NA,
  species_parameters = fgr:::species_parameters,
  fgr_constants = fgr:::fgr_constants,
  moe = NA,
  mor = NA,
  fknot = NA,
  stem_vol = NA,
  crown_vol = NA,
  stem_density = NA,
  crown_density = NA,
  c_reg = NA,
  c_drag = NA,
  n_drag = NA,
  drag_upper_limit = NA,
  snow_depth = NA,
  snow_density = NA,
  ro = NA,
  x = NA,
  y = NA,
  z = NA,
  dams = NA
)
}
\arguments{
\item{stand_id}{Stand identifier. Essential.}

\item{tree_id}{Tree identifier. Essential.}

\item{date}{Date of wind damage risk assessment. Optional (but useful!).}

\item{species}{Tree species under investigation. Essential.}

\item{tree_ht}{Individual tree height (m). Essential.}

\item{dbh}{Diameter of the stem at breast height, i.e. 1.3m above the ground (cm). Essential.}

\item{spacing_current}{Current mean spacing of trees in the stand (m). Essential.}

\item{predominant_species}{Predominant species in the stand. Defaults to \code{species}.}

\item{stand_mean_ht}{Arithmetic mean height of the trees in the stand (m). Essential.}

\item{stand_mean_dbh}{Mean dbh of all trees in the stand (cm) - arithmetic average. Dbh is diameter at breast height, measured at 1.3m above the ground. Essential.}

\item{stand_top_ht}{Height of the dominant tree(s) in the stand (m). Essential.}

\item{full_output}{Switch between full and basic outputs.}

\item{weib_a}{Scale parameter of the Weibull distribution for local wind speeds.}

\item{weib_k}{Shape parameter of the Weibull distribution for local wind speeds.}

\item{ci}{Competition Index (\code{BAL}, \code{Hegyi}, \code{None}) used.}

\item{ci_value}{Value of \code{ci}.}

\item{elev_ht}{The height at which the elevated critical wind speed is calculated (m). For the TMC method, this defaults to 1.05 * stand top height}

\item{cr_width}{Width of the tree crown (m).}

\item{cr_depth}{Length of the tree crown (m).}

\item{stand_cr_width}{Width of the crown of the mean tree in the stand (m).}

\item{stand_cr_depth}{Length of the crown of the mean tree in the stand (m).}

\item{soil_group}{Soil group identifier (1 = freely draining mineral soils; 2 = gleyed mineral soils; 3 = peaty mineral soils; 4 = deep peats).}

\item{rooting}{Rooting depth class (1 = Shallow, 2 = Deep, 3 = Medium).}

\item{dist_edge}{Distance of tree from the upwind edge (m).}

\item{gap_size}{Length of the upwind gap (m).}

\item{spacing_before}{Mean spacing of trees in the stand before any thinning (m).}

\item{years_since_thin}{Number of years since most recent thinning.}

\item{species_parameters}{Data.frame covering all species-specific parameters. Defaults to fgr Internal Data species paramters data.frame. If specified by the user, it can have as many or as few species (i.e. rows) as required. It is recommended that when working with user-defined species, stem volume is specified and provided as an input (see below in input list). When this is not possible, stem volume is calculated with Eq. 4 of Fonweban et al. 2012. In order to allow this, letter codes for User-defined species must begin with a capital "U" for the model to recognise that the Fonweban et al. formula needs applied to calculate stem volume.}

\item{fgr_constants}{Data.frame including all constants required for model execution. Defaults to fgr Internal Data fgr constants data.frame. these include entities that can be individually supplied to the model (e.g. ro, the air density, and snow density - see below in input list), or entities that qould not normally require modification (e.g. constants used in the calculation of the Deflection Loading Factor, the Annual Exceedance Probability, the Edge, Gap, and Gust Factors, aerodynamic quantities such as zpd, z0, and gammasolved - but even von Karman's and the gravitational constants can be modified!)}

\item{moe}{Modulus of Elasticity of green wood (MPa). Advanced Input.}

\item{mor}{Modulus of Rupture of green wood (MPa). Advanced Input.}

\item{fknot}{Knot factor (dimensionless). Advanced Input.}

\item{stem_vol}{Individual tree stem volume (m3). Advanced Input.}

\item{crown_vol}{Volume of the crown of individual trees (m3). Advanced Input.}

\item{stem_density}{Density of green wood of the stem (kg m-3). Advanced Input.}

\item{crown_density}{Density of the crown of individual trees (kg m-3). Advanced Input.}

\item{c_reg}{Regression coefficient of uprooting moment against stem weight (N m kg-1). Advanced Input.}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless). Advanced Input.}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless). Advanced Input.}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1). Advanced Input.}

\item{snow_depth}{Depth of layer of snow on tree crown (cm). Advanced Input.}

\item{snow_density}{Density of snow (kg m-3). Advanced Input.}

\item{ro}{Air density (kg m-3). Advanced Input.}

\item{x}{Spatial coordinate.}

\item{y}{Spatial coordinate.}

\item{z}{Spatial coordinate.}

\item{dams}{DAMS (Detailed Aspect Method of Scoring) value to describe the local wind climate (for UK conditions only).}
}
\value{
If \code{full_output} = 1, a comprehensive list including: stand and tree id's; for breakage and overturning, the critical wind speeds (m s-1) at
 canopy top height and anemometer's height; zero plane displacement height (m), canopy surface roughness (m); canopy drag; ratio of critical wind speed at
 canopy top over friction velocity (uh / u*); drag per unit ground area; the turning moment coefficient; the turning moment ratio; the deflection loading
 factor; the critical breaking moment and the critical overturning moment; the combined effect of edge and gap on the applied bending moment; the probabilities
 of damage; the mode of damage; a summary of the inputs. If \code{full_output} = 0, a much shorter list including: stand id and tree id's; the critical
 wind speeds of breakage and overturning, and the associated probabilities.
}
\description{
Wrapper for the single-tree (turning moment coefficient) method.
}
