% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammasolved_fun.R
\name{gammasolved_fun}
\alias{gammasolved_fun}
\title{Gamma Solved Function.}
\usage{
gammasolved_fun(
  cr_width,
  cr_depth,
  spacing,
  uguess,
  n_drag,
  c_drag,
  drag_upper_limit,
  fgr_constants
)
}
\arguments{
\item{cr_width}{Width of the canopy in windless conditions (m).}

\item{cr_depth}{Length of the canopy in windless conditions (m).}

\item{spacing}{Mean distance between trees (m).}

\item{uguess}{Critical wind speed at canopy top calculated with the roughness or single-tree method (m s-1).}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}
}
\value{
\code{gammasolved}, the ratio of critical wind speed at canopy top over friction velocity (uh / u*).
}
\description{
Calculates the ratio between the critical wind speed at canopy top and the friction velocity (uh / u*).
}
