% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_gap_factor.R
\name{max_gap_factor_fun}
\alias{max_gap_factor_fun}
\title{Maximum Gap Factor}
\usage{
max_gap_factor_fun(gap_size, ht)
}
\arguments{
\item{gap_size}{Length of the upwind gap (m).}

\item{ht}{Height of the tree. In the 'roughness' method, this is stand mean height. In the TMC method, this is 'equivalent mean height' (m).}
}
\value{
\code{max_gap_factor}, the effect of gap alone on the maximum bending moment exerted by the wind on a tree.
}
\description{
Effect of gap size on maximum bending moment
}
