% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_vol_switch.R
\name{stem_vol_fun}
\alias{stem_vol_fun}
\title{Stem Volume Function.}
\usage{
stem_vol_fun(species, dbh, ht)
}
\arguments{
\item{species}{Tree species.}

\item{dbh}{The diameter (cm) of the stem at breast height.}

\item{ht}{The height of the tree.}
}
\value{
The volume (m3) of the stem.
}
\description{
Acts as a switch to select the appropriate function to calculate the stem volume of a tree.
}
