% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_vol_functions.R
\name{stem_volume_fonweban}
\alias{stem_volume_fonweban}
\alias{stem_volume_quebec}
\alias{stem_volume_laasasenaho}
\alias{stem_volume_japanese_larch_japan}
\title{Stem Volume Function.}
\usage{
stem_volume_fonweban(dbh, ht, species, species_parameters)

stem_volume_quebec(dbh, ht, species)

stem_volume_laasasenaho(dbh, ht, species)

stem_volume_japanese_larch_japan(dbh, ht)
}
\arguments{
\item{dbh}{Diameter of the stem at breast height, i.e. 1.3m above the ground (cm). Depending on the method used ('roughness' or 'TMC') this can be either the arithmetic average of the dbh of all the trees in the stand, or the dbh of an individual tree.}

\item{ht}{Tree height. Depending on the method used ('roughness' or 'TMC'), this can be either the mean tree in the stand, or each individual tree (m).}

\item{species}{Tree species.}
}
\value{
The volume of the stem (m3). Applies the method in Fonweban et al. (2012) to calculate stem volume.
}
\description{
Calculates the stem volume of a tree.
}
\section{Functions}{
\itemize{
\item \code{stem_volume_quebec}: Applies the method in Honer (1967) to calculate stem volume.

\item \code{stem_volume_laasasenaho}: Applies the method in Laasasenaho (1982) to calculate stem volume.

\item \code{stem_volume_japanese_larch_japan}: Calculates stem volume for Japanese Larch grown in Japan.
}}

