% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_dimensions.R
\name{Tree_Dimensions_Functions}
\alias{Tree_Dimensions_Functions}
\alias{top_ht_to_mean_ht}
\alias{mean_ht_to_top_ht}
\alias{canopy_width_fun}
\alias{canopy_depth_fun}
\alias{diam_base_fun}
\title{Tree Dimensions Functions}
\usage{
top_ht_to_mean_ht(param0_height, param1_height, top_ht)

mean_ht_to_top_ht(param0_height, param1_height, mean_ht)

canopy_width_fun(param0_cr_width, param1_cr_width, dbh)

canopy_depth_fun(param0_cr_depth, param1_cr_depth, ht)

diam_base_fun(dbh, ht, cr_depth)
}
\arguments{
\item{param0_height}{Species-specific parameter for top/mean tree height calculation.}

\item{param1_height}{Species-specific parameter for top/mean tree height calculation.}

\item{top_ht}{Dominant tree height in the stand.}

\item{mean_ht}{Mean tree height in the stand.}

\item{param0_cr_width}{Species-specific parameter for crown width calculation.}

\item{param1_cr_width}{Species-specific parameter for crown width calculation.}

\item{dbh}{The dbh (cm) of a tree.}

\item{param0_cr_depth}{Species-specific parameter for crown length calculation.}

\item{param1_cr_depth}{Species-specific parameter for crown length calculation.}

\item{ht}{The height (m) of a tree.}

\item{cr_depth}{The length (m) of the tree crown.}
}
\description{
Convert top to mean tree height (and viceversa) and calculate crown dimensions and diameter at base of the tree.
}
