% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turning_Coefficient_Functions.R
\name{Turning_Moment_Coefficient_Functions}
\alias{Turning_Moment_Coefficient_Functions}
\alias{tc_intercept_fun}
\alias{tc_zero_intercept_fun}
\alias{tc_zero_intercept_new_bal_fun}
\alias{tc_zero_intercept_fun_balBA}
\title{Turning Moment Coefficient Functions}
\usage{
tc_intercept_fun(dbh, ht, ci, ci_value)

tc_zero_intercept_fun(dbh, ht, ci, ci_value)

tc_zero_intercept_new_bal_fun(dbh, ht, ci, ci_value)

tc_zero_intercept_fun_balBA(dbh, ht, ci, ci_value)
}
\arguments{
\item{dbh}{Diameter of the stem at breast height, i.e. 1.3m above the ground (cm).}

\item{ht}{Individual tree height (m).}

\item{ci}{Competition Index (\code{bal}, \code{heg}, \code{none}) used.}

\item{ci_value}{Value of \code{ci}.}
}
\description{
Calculate the turning moment coefficient.
}
