% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turning_Moment_Ratios.R
\name{Turning_Moment_Ratios}
\alias{Turning_Moment_Ratios}
\alias{tm_ratio_simple}
\alias{tm_ratio_simple_ci}
\alias{tm_ratio}
\alias{tm_ratio_ci}
\title{Turning Moment Ratio Functions}
\usage{
tm_ratio_simple(spacing_before, spacing_current, years_since_thin)

tm_ratio_simple_ci(spacing_before, spacing_current, years_since_thin, ci)

tm_ratio(spacing_before, spacing_current, years_since_thin, cr_width,
  cr_depth, uguess, n_drag, c_drag, drag_upper_limit, ht, fgr_constants)

tm_ratio_ci(spacing_before, spacing_current, years_since_thin, cr_width,
  cr_depth, uguess, n_drag, c_drag, drag_upper_limit, ht, ci,
  fgr_constants)
}
\arguments{
\item{spacing_before}{Mean spacing of trees in the stand before any thinning (m).}

\item{spacing_current}{Current mean spacing of trees in the stand (m).}

\item{years_since_thin}{Number of years after the latest thinning.}

\item{ci}{Competition Index (\code{bal}, \code{heg}, \code{none}) used.}

\item{cr_width}{Width of the crown of the "equivalent mean tree" in the stand (m).}

\item{cr_depth}{Length of the crown of the "equivalent mean tree" in the stand (m).}

\item{uguess}{Critical wind speed at canopy top calculated with the roughness or single-tree method (m s-1).}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}

\item{ht}{Equivalent mean stand height: the level in the stand responsible for most of the momentum absorption (m).}
}
\description{
Calculate the ratio between the turning moment coefficient before and after thinning in the stand.
}
