% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws_rou_breakage.R
\name{uh_breakage_rou}
\alias{uh_breakage_rou}
\title{Critical wind speed for breakage - Roughness method.}
\usage{
uh_breakage_rou(mean_ht, mean_dbh, spacing, dist_edge, gap_size,
  mean_cr_width, mean_cr_depth, moe, mor, fknot, n_drag, c_drag,
  drag_upper_limit, stem_vol, stem_density, crown_density, snow_depth,
  snow_density, ro, fgr_constants)
}
\arguments{
\item{mean_ht}{Arithmetic mean height of the trees in the stand (m).}

\item{mean_dbh}{Mean dbh of all trees in the stand (cm). Dbh is diameter at breast height, measured at 1.3m above the ground. Essential.}

\item{spacing}{Mean distance between trees (m).}

\item{dist_edge}{Distance from the upwind edge (m).}

\item{gap_size}{Length of the upwind gap (m).}

\item{mean_cr_width}{Width of the crown of the mean tree in the stand (m).}

\item{mean_cr_depth}{Depth of the crown of the mean tree in the stand (m).}

\item{moe}{Modulus of Elasticity of green wood (MPa).}

\item{mor}{Modulus of Rupture of green wood (MPa).}

\item{fknot}{Knot factor (dimensionless).}

\item{n_drag}{N parameter of the drag coefficient formula (dimensionless).}

\item{c_drag}{C parameter of the drag coefficient formula (dimensionless).}

\item{drag_upper_limit}{Maximum wind speed used during the experiments from which \code{n_drag} and \code{c_drag} were derived (m*s-1).}

\item{stem_vol}{Stem volume of the mean tree in the stand (m^3).}

\item{stem_density}{Density of green wood of the stem (kg m-3).}

\item{crown_density}{Crown density of the mean tree in the stand (kg m-3).}

\item{snow_depth}{Depth of layer of snow on tree crown (cm).}

\item{snow_density}{Density of snow (kg m-3).}

\item{ro}{Air density (kg m-3).}
}
\value{
A list including: The critical wind speed (m s-1) for breakage at zero plane displacement height for the roughness method;
The applied bending moment; The zero plane displacement height; The ratio of critical wind speed at canopy top over friction velocity
(uh / u*); The deflection loading factor; The critical breaking moment; The combined effect of edge, gap, and gust on the applied
bending moment.
}
\description{
Calculates the critical wind speed for breakage using the roughness method.
}
